+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - chore: lint example test [#656](https://github.com/hypermodeinc/modus/pull/656)
 - fix: unused imports should not be included in metadata [#657](https://github.com/hypermodeinc/modus/pull/657)
 - fix: improve Go version handling [#660](https://github.com/hypermodeinc/modus/pull/660)
+- fix: update runtime wasm tests [#661](https://github.com/hypermodeinc/modus/pull/661)
 
 ## 2024-12-13 - Runtime 0.15.0