+++ CHANGELOG.md
@@ -1,14 +1,29 @@
 # Change Log
 
-## UNRELEASED
+## 2024-12-23 - Runtime 0.16.0
+
+- fix: unused imports should not be included in metadata [#657](https://github.com/hypermodeinc/modus/pull/657)
+- fix: update runtime wasm tests [#661](https://github.com/hypermodeinc/modus/pull/661)
+- fix: support TinyGo 0.35.0 [#662](https://github.com/hypermodeinc/modus/pull/662)
+- feat: add APIs for local time and time zones [#663](https://github.com/hypermodeinc/modus/pull/663)
+
+## 2024-12-23 - Go SDK 0.16.0
 
 - chore: lint example test [#656](https://github.com/hypermodeinc/modus/pull/656)
 - fix: unused imports should not be included in metadata [#657](https://github.com/hypermodeinc/modus/pull/657)
-- test: add tests for AssemblyScript SDK Transform [#659](https://github.com/hypermodeinc/modus/pull/659)
 - fix: improve Go version handling [#660](https://github.com/hypermodeinc/modus/pull/660)
 - fix: update runtime wasm tests [#661](https://github.com/hypermodeinc/modus/pull/661)
 - fix: support TinyGo 0.35.0 [#662](https://github.com/hypermodeinc/modus/pull/662)
 - feat: add APIs for local time and time zones [#663](https://github.com/hypermodeinc/modus/pull/663)
+
+## 2024-12-23 - AssemblyScript SDK 0.16.0
+
+- test: add tests for AssemblyScript SDK Transform [#659](https://github.com/hypermodeinc/modus/pull/659)
+- feat: add APIs for local time and time zones [#663](https://github.com/hypermodeinc/modus/pull/663)
+
+## 2024-12-23 - CLI SDK 0.16.0
+
+- fix: improve Go version handling [#660](https://github.com/hypermodeinc/modus/pull/660)
 
 ## 2024-12-13 - Runtime 0.15.0